+++ src/api/AnalyticsApi.ts
@@ -29,6 +29,8 @@ export default class AnalyticsApi {
   }
 
   pushAnalyticsLoadEvent(params: SQHDetails) {
+    if (!params.externalUserId || !params.externalAccountId) return;
+
     const tenantAlias = encodeURIComponent(params.tenantAlias);
     const accountId = encodeURIComponent(params.externalAccountId);
     const userId = encodeURIComponent(params.externalUserId);