+++ src/widgets/Widget.ts
@@ -93,7 +93,7 @@ export default abstract class Widget {
 
     this.analyticsApi
       .pushAnalyticsLoadEvent(params)
-      .then((response) => {
+      ?.then((response) => {
         _log(`${params.engagementMedium} loaded event recorded.`);
       })
       .catch((ex) => {
@@ -225,8 +225,10 @@ export default abstract class Widget {
         email: email || null,
         firstName: firstName || null,
         lastName: lastName || null,
-        id: this.context.user.id,
-        accountId: this.context.user.accountId,
+
+        // FIXME: Double check this
+        id: this.context.user!.id,
+        accountId: this.context.user!.accountId,
       };
 
       response = this.widgetApi.upsertUser({
@@ -235,18 +237,12 @@ export default abstract class Widget {
         widgetType: this.type,
         jwt,
       });
-    } else if (this.context.type === "cookie") {
-      let userObj = {
-        email: email || null,
-        firstName: firstName || null,
-        lastName: lastName || null,
-      };
-
-      response = this.widgetApi.cookieUser({
-        user: userObj,
+    } else if (this.context.type === "passwordless") {
+      response = this.widgetApi.render({
+        user: undefined,
         engagementMedium,
         widgetType: this.type,
-        jwt,
+        jwt: undefined,
       });
     } else {
       throw new Error("can't reload an error widget");