+++ src/widgets/Widgets.ts
@@ -6,9 +6,13 @@ import EmbedWidget from "./EmbedWidget";
 import PopupWidget from "./PopupWidget";
 import CtaWidget from "./CtaWidget";
 import Widget, { Params } from "./Widget";
-import { WidgetResult, WidgetContext } from "../types";
+import { WidgetResult, WidgetContext, WithRequired } from "../types";
 import { ConfigOptions, EngagementMedium, WidgetConfig } from "../types";
-import { validateConfig, validateWidgetConfig } from "../utils/validate";
+import {
+  validateConfig,
+  validatePasswordlessConfig,
+  validateWidgetConfig,
+} from "../utils/validate";
 
 const _log = debug("squatch-js:widgets");
 
@@ -56,38 +60,6 @@ export default class Widgets {
   }
 
   /**
-   * This function calls the {@link WidgetApi.cookieUser} method, and it renders
-   * the widget if it is successful. Otherwise it shows the "error" widget.
-   *
-   * @param {Object} config Config details
-   * @param {WidgetType} config.widgetType The content of the widget.
-   * @param {EngagementMedium} config.engagementMedium How to display the widget.
-   * @param {User} config.user An optional user to include
-   * @param {string} config.jwt the JSON Web Token (JWT) that is used to
-   *                            validate the data (can be disabled)
-   *
-   * @return {Promise<WidgetResult>} json object if true, with a Widget and user details.
-   */
-  async createCookieUser(config: WidgetConfig): Promise<WidgetResult> {
-    try {
-      const response = await this.api.cookieUser(config);
-      return {
-        widget: this._renderWidget(response, config, {
-          type: "cookie",
-          engagementMedium: config.engagementMedium,
-        }),
-        user: response.user,
-      };
-    } catch (err) {
-      _log(err);
-      if (err.apiErrorCode) {
-        this._renderErrorWidget(err, config.engagementMedium);
-      }
-      throw err;
-    }
-  }
-
-  /**
    * This function calls the {@link WidgetApi.upsertUser} method, and it renders
    * the widget if it is successful. Otherwise it shows the "error" widget.
    *
@@ -103,15 +75,18 @@ export default class Widgets {
    *
    * @return {Promise<WidgetResult>} json object if true, with a Widget and user details.
    */
-  async upsertUser(config: WidgetConfig) {
+  async upsertUser(config: WithRequired<WidgetConfig, "user">) {
     const raw = config as unknown;
-    const clean = validateWidgetConfig(raw);
+    const clean = validateWidgetConfig(raw) as WithRequired<
+      WidgetConfig,
+      "user"
+    >;
     try {
       const response = await this.api.upsertUser(clean);
       return {
         widget: this._renderWidget(response, clean, {
           type: "upsert",
-          user: clean.user,
+          user: clean.user || null,
           engagementMedium: config.engagementMedium,
           container: config.container,
           trigger: config.trigger,
@@ -142,14 +117,16 @@ export default class Widgets {
    *
    * @return {Promise<WidgetResult>} json object if true, with a Widget and user details.
    */
-  async render(config: WidgetConfig): Promise<WidgetResult> {
+  async render(config: WidgetConfig): Promise<WidgetResult | undefined> {
     const raw = config as unknown;
-    const clean = validateWidgetConfig(raw);
+    const clean = validatePasswordlessConfig(raw);
+
     try {
-      const response = await this.api.cookieUser(clean);
+      const response = await this.api.render(clean);
+
       return {
-        widget: this._renderWidget({ template: response }, clean, {
-          type: "cookie",
+        widget: this._renderWidget(response, clean, {
+          type: "passwordless",
           engagementMedium: clean.engagementMedium,
         }),
         user: response.user,
@@ -234,7 +211,7 @@ export default class Widgets {
     if (!response) throw new Error("Unable to get a response");
 
     let widget;
-    let displayOnLoad = false;
+    let displayOnLoad = !!config.displayOnLoad;
     let displayCTA = false;
     const opts = response.jsOptions || "";